+++ changelog.md
@@ -1,3 +1,8 @@
+## WAP-03
+
+- Ajout de la route /film/{id}
+- Ajout de la page de film
+
 ## WAP-02
 
 - Changement de construction de User pour y un LoginController